Split messages into separate module
