+++ src/components/manage/Blocks/MetadataSection/i18n.js
@@ -1,0 +1,34 @@
+import { defineMessages } from 'react-intl';
+
+const messages = defineMessages({
+  fixed: {
+    id: 'Fixed width table cells',
+    defaultMessage: 'Fixed width table cells',
+  },
+  compact: {
+    id: 'Make the table compact',
+    defaultMessage: 'Make the table compact',
+  },
+  basic: {
+    id: 'Reduce complexity',
+    defaultMessage: 'Reduce complexity',
+  },
+  celled: {
+    id: 'Divide each row into separate cells',
+    defaultMessage: 'Divide each row into separate cells',
+  },
+  inverted: {
+    id: 'Table color inverted',
+    defaultMessage: 'Table color inverted',
+  },
+  striped: {
+    id: 'Stripe alternate rows with color',
+    defaultMessage: 'Stripe alternate rows with color',
+  },
+  defaultFieldset: {
+    id: 'Default',
+    defaultMessage: 'Default',
+  },
+});
+
+export default messages;
